whisper_ecapa_tdnn/whisper_ecapa_tdnn/requirements.txt
openai-whisper==20230314
pyannote.audio==3.0
tabulate>=0.5.0
numpy>=1.16.2
matplotlib>=2.0
wave==0.0.2
scikit-learn==1.2.2
intervaltree>=3.0.0